refactor: fix circular dep
