+++ packages/app/src/systems/Chains/eth/components/EthAssetCard/index.tsx
@@ -1,0 +1,1 @@
+export * from './EthAssetCard';
