+++ packages/app/src/systems/Chains/eth/components/RemoveAssetDialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './RemoveAssetDialog';
