+++ packages/app/src/systems/Chains/eth/containers/index.tsx
@@ -3,4 +3,3 @@ export * from './TxEthToFuelDialog';
 export * from './TxListItemEthToFuel';
 export * from './EthAssetsDialog';
 export * from './AddAssetFormDialog';
-export * from './RemoveAssetDialog';
